+++ Sound.ino
@@ -158,15 +158,15 @@ void EE() {
   Timer3.start(); //Continue responding to the ESP-link website
 }
 
-void buzz(int targetPin, long frequency, long length) {
+void buzz(int targetPin, uint32_t frequency, uint32_t length) {
   digitalWrite(PowerLEDOutPin, HIGH);
-  long delayValue = 1000000 / frequency / 2; // calculate the delay value between transitions
+  uint32_t delayValue = 1000000 / frequency / 2; // calculate the delay value between transitions
   //// 1 second's worth of microseconds, divided by the frequency, then split in half since
   //// there are two phases to each cycle
-  long numCycles = frequency * length / 1000; // calculate the number of cycles for proper timing
+  uint32_t numCycles = frequency * length / 1000; // calculate the number of cycles for proper timing
   //// multiply frequency, which is really cycles per second, by the number of seconds to
   //// get the total number of cycles to produce
-  for (long i = 0; i < numCycles; i++) { // for the calculated length of time...
+  for (uint32_t i = 0; i < numCycles; i++) { // for the calculated length of time...
     digitalWrite(targetPin, HIGH); // write the buzzer pin high to push out the diaphram
     delayMicroseconds(delayValue); // wait for the calculated delay value
     digitalWrite(targetPin, LOW); // write the buzzer pin low to pull back the diaphram
